requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
azure-storage-blob==12.24.0
python-multipart==0.0.20
pydantic[email]==2.10.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.2
jinja2==3.1.5
